better slice construction: now a general function atDepth, where top level is depth 0 etc. also atLevel, so you can get bottom level aswell.
Other changes: 
flattenSTs, when a list of slice transfomrations are flattened into a 
slice, they are applied til a list of All with same length as max depth 
of sts. This means that we can cleanly add deeper levels as we go from 
left to right in prefixtree, and perform functions on these. useful when 
giving a chord a playing style... 

cv with chords basically works